Added new code
Wrote program to code a number in a numerical base that is in a text file will be converted to binary and coded to binary.
Wrote program that converts a binary file to a text file as a number in a numerical base.
Wrote a robot program that see's if a file is a real file. If not it is deleted.