Included pic in top directory. More documentation
02/11/2017
Included a picture showing how to move the destination file up and
down then use the time scanners and difference files generated from
/dev/random or the random number generators or provided by the user
to generate target files. Included the slop directory in tools
to show the user how to move the destination file up and down
to construct target files.