Wrote easy appliction, added firecodes bpchecksums
Einstein said all space-time exist at the same time. So all computer files exist at the same time too. This project maps every computer file in space-time that can be reconstructed on the hard drive. There are always at least two copies of computer files. 1 copy in existence in space-time and 1 copy in the informational universe. When the current copy of that file is not available in the present it can be downloaded from the informational universe to the present. The project allows for computer time travel, military strength data recovery, and archiving of computer files as check-sums.